requirement.txt
aiohttp==3.7.3
Faker==5.1.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
importlib-metadata==3.3.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpath==0.82
multidict==5.1.0
openpyxl==3.0.5
PyMySQL==0.10.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
SQLAlchemy==1.3.20
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
xlrd==2.0.1
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-APScheduler==1.12.2
gunicorn==20.1.0
DingtalkChatbot=1.5.3
simplejson=3.17.6